explore/reader done

+++ client/js/components/postlist.js
@@ -9,22 +9,27 @@ Vue.component('postlist', {
     },
     methods: {
         deletePost(id) {
-            axios({
-                method: 'DELETE',
-                url: serverURL + '/articles/' + id, 
-                headers: this.$root.config.headers,
-            })
-                .then(({data}) => {
-                    this.getPosts()
-                    this.posts = this.posts.filter(x => x._id !== id)
+            if(localStorage.getItem('miniwp_token')) {
+                axios({
+                    method: 'DELETE',
+                    url: serverURL + '/articles/' + id, 
+                    headers: this.$root.config.headers,
                 })
-                .catch(({response}) => {
-                    swal.fire({
-                        title: 'Error!',
-                        text: response.data,
-                        type: 'error',
+                    .then(({data}) => {
+                        this.getPosts()
+                        this.posts = this.posts.filter(x => x._id !== id)
                     })
-                })
+                    .catch(({response}) => {
+                        swal.fire({
+                            title: 'Error!',
+                            text: response.data,
+                            type: 'error',
+                        })
+                    })
+            }
+            else {
+                this.$root.showLogin()
+            }
         },
         updatePost(id) {
             this.$root.showUpdatePost(id)
